ceylon.test: TapLoggingListener: move plan to top
The “plan” is the number of tests you want to run, like `1..10`. The
intention is that even if the test runner crashes, you can still know
how far you got, and how many tests should be considered failed. This,
of course, only works if you print the plan right at the beginning.