idempotent war initialization + expose init method
ServletContextListener.contextInitialized() is called after
ServletContainerInitializer.onStartup(), which is too late if Ceylon is used to
implement a ServletContainerInitializer.

This change makes it possible to register and perform initialization with a
WarInitializer from the ServletContainerInitializer itself, so that all servlet
container initialization tasks can be performed with a fully initialized Ceylon
environment.