add loggerFactory, support logging file & line no.
This commit alters the primary factory method to include an optional
ClassOrInterface parameter to specify a class that will be used to wrap the
logger:

    logger<Wrapper>(Category category,
                    ClassOrInterface<Wrapper>? wrapper = null)

This information is necessary for loggers that include the file and line number
of the originating logging statement. Typically, for Java loggers, this
involves passing the fully qualified class name (fqcn) of the wrapper class
when forwarding log messages.