Fix HashSet rehasing issue
During rehasing, only the last 2 cell would be preserved per bucket
in the new store, leading to elements disappearing during rehashing.
The old code finds the the last cell in the bucket and links it to
the new cell.
Since order in the buckets does not matter (even when stability ==
linked), remove this search and simply insert the new cell at the
head of the bucket.