Add wrapper types for commonly used types to add unique typing.
This is useful in contexts where several Integer variables are used for
different concepts such as AccountNumber, ReferenceNumber, etc.  This
adds more safety than raw [[Integer]]s or [[String]]s, etc, and prevents
accidental equality that would otherwise occur with these types, or
passing such values in the wrong other to functions.

The idea comes from the following article:   

http://techblog.realestate.com.au/the-abject-failure-of-weak-typing/

The implementations of WrapperType include WrappedByte, WrappedBoolean,
WrappedFloat, and WrappedInteger.  In addition, a WrappedTypedComparable
exists as well as its subclass, WrappedString.  These are all meant to
be subclassed by users (ex FirstName as a subclass of WrappedString). It
is possible for implementors to created their own wrapper types.  For
example, a Widget class may be wrapped with a WrappedWidget, which
subclasses WrapperType.