Fix for #583
The problem was twofold.
The first issue was that our milliseconds fraction calculation was
really simplistic and only worked if the milliseconds part was exactly
three digits long.

Second problem was that our fractional part calculation only affected
next smaller unit of time. Thus, 0.025 hours is going to be less than
a minute, thus minutes part remained 0 and rest was fixed to 0s as
well.

The same issue was affecting Time parsing as well