fix linkedResource for circular symlinks for #560
This is a breaking change since Resource.linkedResource now returns
`Resource` instead of File|Directory|Nil. This is necessary, since links
that are part of a cycle don't ultimately resolve to one of the three
previous options!

Note: the previous code returned `Nil` when attempting to obtain a
resource for a cirular symlink, but with a path that actually pointed to
an existing symlink.