Handle out-of-range JSON numbers as Floats for #661
The premise of #661 is that JSON numbers that should be treated as
Ceylon Floats don't always have decimal points. Stated differently, the
absence of '.' in a JSON number is not a reliable indicator of the
desired Ceylon numeric type.

So, with this commit, if we are unable to parse a number as an Integer,
fall back to treating it as a Float.